Update RadioEvent.h
Update type for snr argument to PacketRx

+++ examples/inc/RadioEvent.h
@@ -14,7 +14,7 @@ public:
 
     virtual ~RadioEvent() {}
 
-    virtual void PacketRx(uint8_t port, uint8_t *payload, uint16_t size, int16_t rssi, int8_t snr, lora::DownlinkControl ctrl, uint8_t slot, uint8_t retries, uint32_t address, bool dupRx) {
+    virtual void PacketRx(uint8_t port, uint8_t *payload, uint16_t size, int16_t rssi, int16_t snr, lora::DownlinkControl ctrl, uint8_t slot, uint8_t retries, uint32_t address, bool dupRx) {
         mDotEvent::PacketRx(port, payload, size, rssi, snr, ctrl, slot, retries, address, dupRx);
 
 #if ACTIVE_EXAMPLE == FOTA_EXAMPLE
